Bugfix Update(version 0.8.1.2)
Little changes
