+++ src/main/java/core/logic/ApiConfigs.java
@@ -46,7 +46,7 @@ public class ApiConfigs {
                 String value = jsonResponse.optString(key);
 
                 out.println(alignment(58) + getAnsi256Color(systemLayoutColor) + text + " " + RESET +
-                        getAnsi256Color(systemFirstColor) + value + RESET);
+                        getAnsi256Color(systemMainColor) + value + RESET);
             } catch (Exception e) {
                 message("Error parsing JSON response: " + e.getMessage(), systemRejectionColor,58,0,out::print);
             }
@@ -70,7 +70,7 @@ public class ApiConfigs {
                 double temperature = currentWeather.getDouble("temperature");
 
                 out.println(alignment(58) + getAnsi256Color(systemLayoutColor) + "Weather in " + city + " now: " + RESET
-                        + getAnsi256Color(systemFirstColor) + temperature + "°C" + RESET);
+                        + getAnsi256Color(systemMainColor) + temperature + "°C" + RESET);
             } catch (Exception e) {
                 message("Error parsing JSON response: " + e.getMessage(), systemRejectionColor,58,0,out::print);
             }
