+++ src/main/java/extensions/finance/CryptoConfigs.java
@@ -1,23 +1,11 @@
 package extensions.finance;
 
-import org.json.JSONObject;
-
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
+import lombok.Getter;
 import java.util.*;
 
-import static core.logic.ApiConfigs.httpRequest;
-import static core.logic.AppearanceConfigs.*;
-import static core.logic.CommandManager.*;
-import static core.logic.DisplayManager.displayErrorAscii;
-import static core.logic.TextConfigs.*;
-
-import static java.lang.System.out;
-
-//Not a final version
 public class CryptoConfigs {
 
-    private static final LinkedList<String> cryptocurrencyCodes = new LinkedList<>() {{
+    static final LinkedList<String> cryptocurrencyCodes = new LinkedList<>() {{
         add("btc");
         add("eth");
         add("xrp");
@@ -62,7 +50,7 @@ public class CryptoConfigs {
         add("ape");
     }};
 
-    private static final LinkedList<String> cryptocurrencyName = new LinkedList<>() {{
+    static final LinkedList<String> cryptocurrencyName = new LinkedList<>() {{
         add("bitcoin");
         add("ethereum");
         add("ripple");
@@ -107,6 +95,7 @@ public class CryptoConfigs {
         add("apecoin");
     }};
 
+    @Getter
     public static String[] cryptoLogo = {
             "  .oooooo.                                       .             ",
             " d8P'  `Y8b                                    .o8             ",
@@ -120,247 +109,23 @@ public class CryptoConfigs {
             " "
     };
 
-    private static final Scanner scanner = new Scanner(System.in);
-    private static String cryptocurrencyCode;
-    private static String fiatCurrencyCode;
+    static final Scanner scanner = new Scanner(System.in);
+    static String cryptocurrencyCode;
+    static String fiatCurrencyCode;
 
-    private static int themeColor_1 = 85;
-    private static int themeColor_2 = 79;
-    private static int layoutColor = 15;
+    static int themeColor_1 = 80;
+    static int themeColor_2 = 73;
+    static int layoutColor = 15;
 
-    private static int acceptanceColor = 46;
-    private static int rejectionColor = 196;
+    static int acceptanceColor = 46;
+    static int rejectionColor = 196;
 
-    private static final Map<String, String> CRYPTO_MAP = new HashMap<>() {{
+    @Getter
+    private static final String version = "1.1.3";
+
+    static final Map<String, String> CRYPTO_MAP = new HashMap<>() {{
         for (int i = 0; i < cryptocurrencyCodes.size() && i < cryptocurrencyName.size(); i++) {
             put(cryptocurrencyCodes.get(i), cryptocurrencyName.get(i));
         }
     }};
-
-    //Main menu of extension
-    public static void displayCryptoMenu() {
-        modifyMessage('n', 2);
-        switchLogoRandomly(cryptoLogo,8);
-        marginBorder(1,2);
-
-        displayConfirmation("Enter","to open and","to skip", themeColor_1, rejectionColor, layoutColor);
-        modifyMessage('n',1);
-
-        choice("List of cryptocurrencies", CryptoConfigs::displayListOfCryptocurrencies,
-                themeColor_1, layoutColor, rejectionColor);
-
-        choice("Exchanger", CryptoConfigs::exchanger,
-                themeColor_1, layoutColor, rejectionColor);
-        modifyMessage('n',2);
-
-        choice("Price tracker", CryptoConfigs::currencyPriceTracker,
-                themeColor_1, layoutColor, rejectionColor);
-        marginBorder(2,1);
-    }
-
-    //Exchanger method
-    private static void exchanger() {
-        modifyMessage('n',1);
-        alert("i", "Type '" + getAnsi256Color(themeColor_1) + "exit"
-                + getAnsi256Color(layoutColor) + "' to\n" + alignment(58) + "quit the extension.",
-                58, themeColor_1, layoutColor);
-        modifyMessage('n',1);
-
-        while (true) {
-            modifyMessage('n', 1);
-            out.print(alignment(58) + getAnsi256Color(layoutColor) + "Cryptocurrency code: " + RESET);
-            cryptocurrencyCode = scanner.nextLine().trim().toLowerCase();
-
-            if (cryptocurrencyCode.equalsIgnoreCase("exit")) {
-                modifyMessage('n', 1);
-                terminateExtension(themeColor_1,acceptanceColor,layoutColor);
-                return;
-            }
-
-            if (!CRYPTO_MAP.containsKey(cryptocurrencyCode)) {
-                message("Invalid cryptocurrency code: " + cryptocurrencyCode, rejectionColor, 58, 0, out::print);
-                continue;
-            }
-
-            out.print(alignment(58) + getAnsi256Color(layoutColor) + "Fiat currency code: " + RESET);
-            fiatCurrencyCode = scanner.nextLine().trim().toLowerCase();
-
-            if (fiatCurrencyCode.equalsIgnoreCase("exit")) {
-                terminateExtension(themeColor_1,acceptanceColor,layoutColor);
-                return;
-            }
-
-            if (fiatCurrencyCode.isEmpty()) {
-                message("Fiat currency code cannot be empty.", rejectionColor, 58, 0, out::print);
-                continue;
-            }
-
-            getCryptocurrencyPrice(CRYPTO_MAP.get(cryptocurrencyCode), fiatCurrencyCode);
-        }
-    }
-
-    //Tracker
-    private static void currencyPriceTracker() {
-        modifyMessage('n', 2);
-        alert("i", "Type '" + getAnsi256Color(rejectionColor) + "exit"
-                + getAnsi256Color(layoutColor) + "' to quit the extension at any time.\n"
-                + alignment(58) + "You cannot exit this mode while tracking is in progress.", 58,
-                themeColor_1, layoutColor);
-        while (true) {
-            modifyMessage('n', 1);
-            out.print(alignment(58) + getAnsi256Color(layoutColor) + "Cryptocurrency code: " + RESET);
-            cryptocurrencyCode = scanner.nextLine().trim().toLowerCase();
-
-            if (cryptocurrencyCode.equalsIgnoreCase("exit")) {
-                message("\r   Status: " + getAnsi256Color(acceptanceColor) + "✓", layoutColor,58,0,out::print);
-                message("Extension terminated correctly", themeColor_1,
-                        58,0,out::print);
-                return;
-            }
-
-            if (!CRYPTO_MAP.containsKey(cryptocurrencyCode)) {
-                message("Invalid cryptocurrency code: " + cryptocurrencyCode, rejectionColor, 58, 0, out::print);
-                continue;
-            }
-
-            out.print(alignment(58) + getAnsi256Color(layoutColor) + "Fiat currency code: " + RESET);
-            fiatCurrencyCode = scanner.nextLine().trim().toLowerCase();
-
-            if (fiatCurrencyCode.equalsIgnoreCase("exit")) {
-                message("\r   Status: " + getAnsi256Color(acceptanceColor) + "✓", layoutColor,58,0,out::print);
-                message("Extension terminated correctly", themeColor_1,
-                        58,0,out::print);
-                return;
-            }
-
-            if (fiatCurrencyCode.isEmpty()) {
-                message("Fiat currency code cannot be empty.", rejectionColor, 58, 0, out::print);
-                continue;
-            }
-
-            out.print(alignment(58) + getAnsi256Color(layoutColor) + "Duration in minutes: " + RESET);
-
-            double duration;
-            try {
-                duration = Integer.parseInt(scanner.nextLine().trim()) * 60000;
-            } catch (NumberFormatException e) {
-                modifyMessage('n', 1);
-                message("Invalid duration. Please enter a valid number.", rejectionColor, 58, 0, out::print);
-                continue;
-            }
-
-            if (duration <= 0) {
-                modifyMessage('n', 1);
-                message("Duration must be greater than zero.", rejectionColor, 58, 0, out::print);
-                continue;
-            }
-
-            for (int i = 0; i < duration; i += 5000) {
-                getCryptocurrencyPrice(CRYPTO_MAP.get(cryptocurrencyCode), fiatCurrencyCode);
-                try {
-                    Thread.sleep(40000); //pause 40 sec
-                } catch (InterruptedException e) {
-                    errorFormatting("Tracking interrupted.");
-                    return;
-                }
-            }
-        }
-    }
-
-    //Lists of cryptocurrencies
-    private static void displayListOfCryptocurrencies() {
-        try {
-            if (cryptocurrencyName.size() != cryptocurrencyCodes.size()) {
-                throw new IllegalStateException("Cryptocurrency lists are not synchronized.");
-            }
-
-            int totalItems = cryptocurrencyName.size();
-            int midPoint = (totalItems + 1) / 2;
-
-            List<String> leftColumnNames = cryptocurrencyName.subList(0, midPoint);
-            List<String> leftColumnCodes = cryptocurrencyCodes.subList(0, midPoint);
-            List<String> rightColumnNames = cryptocurrencyName.subList(midPoint, totalItems);
-            List<String> rightColumnCodes = cryptocurrencyCodes.subList(midPoint, totalItems);
-
-            int maxRows = Math.max(leftColumnNames.size(), rightColumnNames.size());
-
-            modifyMessage('n',1);
-            out.printf(alignment(38) + getAnsi256Color(layoutColor) + "%-40s"
-                    + alignment(10) + getAnsi256Color(layoutColor) + "%-40s%n",
-                    getAnsi256Color(themeColor_1) + "Names",getAnsi256Color(themeColor_1) + "          Codes");
-
-            for (int i = 0; i < maxRows; i++) {
-                String leftEntry = i < leftColumnNames.size()
-                        ? "· " + capitalizeMessage(leftColumnNames.get(i))
-                        + getAnsi256Color(themeColor_2)
-                        + " [" + leftColumnCodes.get(i).toUpperCase() + "]"
-                        : "";
-
-                String rightEntry = i < rightColumnNames.size()
-                        ? "· " + capitalizeMessage(rightColumnNames.get(i))
-                        + getAnsi256Color(themeColor_2)
-                        + " [" + rightColumnCodes.get(i).toUpperCase() + "]"
-                        : "";
-
-                out.printf(alignment(38) + getAnsi256Color(layoutColor) + "%-40s"
-                        + alignment(10) + getAnsi256Color(layoutColor) + "          %-40s%n", leftEntry, rightEntry);
-            }
-            marginBorder(2,2);
-        } catch (Exception e) {
-            errorFormatting("Error of showing list");
-        }
-    }
-
-    //Getting an actual cryptocurrency prices
-    private static void getCryptocurrencyPrice(String cryptocurrencyCode, String fiatCurrencyCode) {
-        LocalDateTime localTime = LocalDateTime.now();
-        DateTimeFormatter myFormatter = DateTimeFormatter.ofPattern("HH:mm:ss");
-        String formattedTime = localTime.format(myFormatter);
-
-        String response = httpRequest(
-                "https://api.coingecko.com/api/v3/simple/price?ids="
-                        + cryptocurrencyCode + "&vs_currencies=" + fiatCurrencyCode,
-                "GET", "", "response");
-
-        if (response != null) {
-            try {
-                JSONObject jsonResponse = new JSONObject(response);
-                if (jsonResponse.has(cryptocurrencyCode)) {
-                    double price = jsonResponse.getJSONObject(cryptocurrencyCode).getDouble(fiatCurrencyCode);
-                    out.println(alignment(58) + getAnsi256Color(themeColor_1)
-                            + capitalizeMessage(cryptocurrencyCode) + getAnsi256Color(layoutColor) + " costs in "
-                            + fiatCurrencyCode.toUpperCase() + ": "
-                            + getAnsi256Color(themeColor_1) + price + getAnsi256Color(layoutColor)
-                            + " [" + formattedTime + "]");
-                } else {
-                    errorFormatting("Invalid response from API.");
-                }
-            } catch (Exception e) {
-                errorFormatting("Error parsing response: " + e.getMessage() + ".");
-            }
-        } else {
-            errorFormatting("Failed to fetch price. Check your network connection.");
-        }
-    }
-
-    //Show settings
-   /* private static void extensionSettings(){
-        modifyMessage('n', 2);
-        try{
-
-        }
-        catch (Exception e){
-            errorFormatting("Unexpected error.");
-        }
-    }
-    */
-
-    //Formating the error outputs
-    private static void errorFormatting(String text) {
-        modifyMessage('n', 1);
-        displayErrorAscii();
-        modifyMessage('n',1);
-        message(text, rejectionColor, 58, 0, out::print);
-        modifyMessage('n', 1);
-    }
 }
