--- src/main/java/extensions/finance/CryptoPage.java
@@ -1,10 +1,0 @@
-package extensions.finance;
-
-
-import static extensions.finance.CryptoConfigs.displayCryptoMenu;
-
-public class CryptoPage {
-    public static void exchangerPage() {
-        displayCryptoMenu();
-    }
-}
