--- src/main/java/extensions/math/CalculatorConfig.java
@@ -1,5 +1,0 @@
-package extensions.math;
-
-public class CalculatorConfig {
-
-}
