--- src/main/java/extensions/math/CalculatorPage.java
@@ -1,5 +1,0 @@
-package extensions.math;
-
-public class CalculatorPage {
-
-}
